Add files via upload
This is a simple method #substrings that takes a string as the first argument and  an array of valid substrings (my dictionary) as the second argument. This method return a hash listing each substring (case insensitive) that was found in the original string and how many times it was found.